java: Add FJakartaServlet
The Java ecosystem is switching from javax to jakarta, but it will be a
slow transition.  For now, simply copy FServlet to FJakartaServlet so
that consumers can choose the version that works for their service (the
provided scope used by the new dependency and already used by the
existing javax.servlet-api dependency ensures that consumers don't get
unwanted dependency).